// src/app/modules/domain/modules/clients/client-dialog/typings/client-form-field.ts
export enum ClientFormFields {
    ID = 'id',
    FIRSTNAME = 'firstname',
    SURNAME = 'surname',
    AGE = 'age',
    DRIVER_LICENSE_NUMBER = 'driver_license_number',
    STREET = 'street',
    HOUSE_NUMBER = 'house_number',
    APARTMENT_NUMBER = 'apartment_number',
    CITY = 'city',
    POST_CODE = 'post_code'
  };